+++ TelcobrightVS13/Projects/Models_Mediation/rate.cs
@@ -26,7 +26,7 @@ namespace MediationModel
         public int Resolution { get; set; }
         public float MinDurationSec { get; set; }
         public int SurchargeTime { get; set; }
-        public double SurchargeAmount { get; set; }
+        public decimal SurchargeAmount { get; set; }
         public Nullable<long> idrateplan { get; set; }
         public string CountryCode { get; set; }
         public Nullable<System.DateTime> date1 { get; set; }
@@ -44,14 +44,14 @@ namespace MediationModel
         public Nullable<decimal> OtherAmount1 { get; set; }
         public Nullable<decimal> OtherAmount2 { get; set; }
         public Nullable<decimal> OtherAmount3 { get; set; }
-        public Nullable<double> OtherAmount4 { get; set; }
-        public Nullable<double> OtherAmount5 { get; set; }
-        public Nullable<double> OtherAmount6 { get; set; }
+        public Nullable<decimal> OtherAmount4 { get; set; }
+        public Nullable<decimal> OtherAmount5 { get; set; }
+        public Nullable<decimal> OtherAmount6 { get; set; }
         public Nullable<float> OtherAmount7 { get; set; }
         public Nullable<float> OtherAmount8 { get; set; }
         public Nullable<float> OtherAmount9 { get; set; }
         public Nullable<float> OtherAmount10 { get; set; }
-        public double TimeZoneOffsetSec { get; set; }
+        public decimal TimeZoneOffsetSec { get; set; }
         public Nullable<int> RatePosition { get; set; }
         public Nullable<float> IgwPercentageIn { get; set; }
         public string ConflictingRateIds { get; set; }
